% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalCCC.utils.R
\name{causalCCC.mosaic}
\alias{causalCCC.mosaic}
\title{Create Mosaic datatable for causalCCC}
\usage{
causalCCC.mosaic(
  data_input,
  assay_name = "RNA",
  interact_ident,
  senders_name,
  receivers_name,
  genes_senders,
  genes_receivers,
  metadata_senders = NULL,
  metadata_receivers = NULL
)
}
\arguments{
\item{data_input}{[a dataframe or a Seurat object]
A Single-Cell transcriptomics object, dataframe or Seurat. If
dataframe must contains genes and metadata as variables and
cells as observations.}

\item{assay_name}{[a string]. Gives the name of the assay to take
the transcriptomics raw counts from (usually 'RNA')}

\item{interact_ident}{[a string]. Gives the name of the metadata
containing the celltypes population}

\item{senders_name}{[a string]. Gives the name of the senders}

\item{receivers_name}{[a string]. Gives the name of the receivers}

\item{genes_senders}{[a vector] A list of selected genes (strings) for the
senders cells. Overlap with genes_receivers is not an issue}

\item{genes_receivers}{[a vector] A list of selected genes (strings) for the
receivers cells. Overlap with genes_senders is not an issue}

\item{metadata_senders}{[a vector] A list of metadata (strings) that will
be added to the genes network. They can be relevant to senders
and/or receivers cells.}

\item{metadata_receivers}{[a vector] A list of metadata (strings) that will
be added to the genes network. They can be relevant to senders
and/or receivers cells.}
}
\value{
A mosaic datatable that can be used for causalCCC
network reconstruction.
}
\description{
Function that takes an data_input raw counts,
 senders and receivers information metadata and create a mosaic
 datable that contains first senders network
 variables then receivers network variables
}
