% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalCCC.wrapper.R
\name{causalCCC.wrapper}
\alias{causalCCC.wrapper}
\title{Run the preprocessing of an Seurat object for causalCCC}
\usage{
causalCCC.wrapper(
  data_input,
  assay_name = "RNA",
  interact_ident = NULL,
  senders_name = NULL,
  receivers_name = NULL,
  do_CCC = FALSE,
  CCC_method = "cellphonedb",
  n_CCClinks = 20,
  interact_edges = NULL,
  do_MIselect = FALSE,
  n_genes_senders = NULL,
  n_genes_receivers = NULL,
  save = TRUE,
  plot = TRUE,
  goi_senders = NULL,
  goi_receivers = NULL,
  metadata_senders = list(),
  metadata_receivers = list(),
  species = "human",
  genes_senders = NULL,
  genes_receivers = NULL,
  do_layout = FALSE,
  wd_path = getwd()
)
}
\arguments{
\item{data_input}{[a dataframe or a Seurat object]
A Single-Cell transcriptomics object, dataframe or Seurat. If
dataframe must contains genes and metadata as variables and
cells as observations.}

\item{assay_name}{[a string]. Gives the name of the assay to take
the transcriptomics raw counts from (usually 'RNA')}

\item{interact_ident}{[a string]. Gives the name of the metadata
containing the celltypes population}

\item{senders_name}{[a string]. Gives the name of the senders population.}

\item{receivers_name}{[a string]. Gives the name of the receivers population.}

\item{do_CCC}{A boolean. If TRUE, perform Cell-Cell Communication analysis.}

\item{CCC_method}{A character string specifying the CCC method to use, see 
show_methods() of LIANA}

\item{n_CCClinks}{A numeric specifiying how many L-R to keep. Default is 60.}

\item{do_MIselect}{A boolean. If TRUE, perform Mutual Information-based feature selection.}

\item{n_genes_senders}{[a numeric] The number of genes to keep after pairwise
mutual information ranking.  Default depends on the length of goi
(between 15 and 100).}

\item{n_genes_receivers}{[a numeric] The number of genes to keep after pairwise
mutual information ranking.  Default depends on the length of goi
(between 15 and 100).}

\item{save}{A boolean. If TRUE, save the MI table of MIselection to wd_path.}

\item{plot}{A boolean. If TRUE, generate MI heatmaps plots and save them in wd_path.}

\item{goi_senders}{A vector. A list of genes of interest for the senders cells.}

\item{goi_receivers}{A vector. A list of genes of interest for the receivers cells.}

\item{metadata_senders}{[a named list] A named list of metadata with their levels
that will be added to the genes network. They are relevant to senders. See exemple}

\item{metadata_receivers}{[a named list] A named list of metadata with their levels
that will be added to the genes network. They are relevant to receivers. See exemple}

\item{species}{A string. The species ('mouse' or 'human').}

\item{genes_senders}{[a vector] An optional list of genes to add to the senders network.}

\item{genes_receivers}{[a vector] An optional list of genes to add to the receivers network.}

\item{do_layout}{A boolean. If TRUE, create a network layout.json file for causalCCC web server.}

\item{wd_path}{A string. The working directory path for saving results.}

\item{ligands}{A vector. An optional list of ligand genes to add to the network.}

\item{receptors}{A vector. An optional list of receptor genes to add to the network.}
}
\value{
A list containing mosaic datatable, state order, interact edges, 
and network layout for causalCCC network reconstruction.
}
\description{
Function that takes an data_input raw counts,
 senders and receivers information metadata and create the
output files necessary to run causalCCC
}
\examples{
\donttest{
metadata <- list(
 treatment = c("Control", "Treated"),
 another_meta = c("Level1", "Level2", "Level3"),
 continusous_meta = NULL
)
}

}
